+++ src/seqgen.cpp
@@ -47,7 +47,7 @@ void usage() {
 
 int main(int argc, char **argv) {
 
-  unsigned int seqlen{0};
+  unsigned long seqlen{0};
   int opt;
   size_t alphlen;
   ofstream outfile;
@@ -78,7 +78,7 @@ int main(int argc, char **argv) {
                   has_out = true;
                 }
                 break;
-      case 'l': if (optarg) seqlen = atoi(optarg);
+      case 'l': if (optarg) seqlen = atol(optarg);
                 break;
       case 's': if (optarg) iseed = atoi(optarg);
                 break;
@@ -153,13 +153,13 @@ int main(int argc, char **argv) {
   if (!has_freqs) {
 
     if (has_out) {
-      for (unsigned int i = 0; i < seqlen; ++i) {
+      for (unsigned long i = 0; i < seqlen; ++i) {
         outfile << lets[gen() % alphlen];
       }
       outfile << '\n';
       outfile.close();
     } else {
-      for (unsigned int i = 0; i < seqlen; ++i) {
+      for (unsigned long i = 0; i < seqlen; ++i) {
         cout << lets[gen() % alphlen];
       }
       cout << '\n';
@@ -170,13 +170,13 @@ int main(int argc, char **argv) {
     discrete_distribution<unsigned int> next_let(freqs.begin(), freqs.end());
 
     if (has_out) {
-      for (unsigned int i = 0; i < seqlen; ++i) {
+      for (unsigned long i = 0; i < seqlen; ++i) {
         outfile << lets[next_let(gen)];
       }
       outfile << '\n';
       outfile.close();
     } else {
-      for (unsigned int i = 0; i < seqlen; ++i) {
+      for (unsigned long i = 0; i < seqlen; ++i) {
         cout << lets[next_let(gen)];
       }
       cout << '\n';
